pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gatomia"
version = "1.0.3"
description = "Transform codebases into comprehensive documentation using AI-powered analysis"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "GatomIA Contributors" }]
keywords = ["documentation", "code-analysis", "ai", "llm", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.1.0",
    "keyring>=24.0.0",
    "GitPython>=3.1.40",
    "Jinja2>=3.1.6",
    "tree-sitter>=0.23.2",
    "tree-sitter-language-pack>=0.8.0",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-javascript>=0.21.4",
    "tree-sitter-typescript>=0.21.2",
    "tree-sitter-c>=0.21.4",
    "tree-sitter-cpp>=0.23.4",
    "tree-sitter-c-sharp>=0.23.1",
    "tree-sitter-php>=0.23.0",
    "openai>=1.107.0",
    "litellm>=1.81.5",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pydantic-ai>=1.0.6",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "networkx>=3.6.0",
    "psutil>=7.2.0",
    "PyYAML>=6.0.2",
    "mermaid-parser-py>=0.0.2",
    "mermaid-py>=0.8.3",
    "github-copilot-sdk>=0.1.0",
    "anthropic>=0.77.0",
    "claude-agent-sdk>=0.1.26",
]

[external]
# Node.js is required for mermaid-py which validates mermaid diagrams in generated documentation
build-requires = [{ name = "nodejs", version = ">=14.0.0" }]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
mia = "gatomia.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/gatomia"
Documentation = "https://github.com/yourusername/gatomia/docs"
Repository = "https://github.com/yourusername/gatomia"
Issues = "https://github.com/yourusername/gatomia/issues"

[tool.setuptools]
packages = [
    "gatomia",
    "gatomia.cli",
    "gatomia.cli.commands",
    "gatomia.cli.models",
    "gatomia.cli.utils",
    "gatomia.cli.adapters",
    "gatomia.src",
    "gatomia.src.be",
    "gatomia.src.be.agent_tools",
    "gatomia.src.be.dependency_analyzer",
    "gatomia.src.be.dependency_analyzer.analysis",
    "gatomia.src.be.dependency_analyzer.analyzers",
    "gatomia.src.be.dependency_analyzer.models",
    "gatomia.src.be.dependency_analyzer.utils",
    "gatomia.src.fe",
]

[tool.setuptools.package-data]
gatomia = ["templates/**/*", "py.typed"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=gatomia --cov-report=term-missing"
